refactor(api): neutralize billing restore and teams connect service naming
Rename the billing-restore and teams-connect services to domain-neutral
filenames and drop the redundant brand prefix from the one branded export
in the teams-connect service. Phase 2 naming cleanup.

- services/zero-billing-restore.service.ts -> services/billing-restore.service.ts
- services/zero-teams-connect.service.ts -> services/teams-connect.service.ts
- zeroTeamsConnectStatus -> teamsConnectStatus

The billing-restore rename is content-free; all of its exports were already
neutral. The agent-compose consolidation preflight manifests track these
service paths as scanned literals, so their entries move to the new sorted
positions to keep test:migration-consistency green.

Closes #27919
