+++ turbo/apps/api/src/signals/routes/billing-restore.ts
@@ -13,7 +13,7 @@ import { organizationAuthContext$ } from "../auth/auth-context";
 import { authRoute } from "../auth/auth-route";
 import { publicBrand$ } from "../context/hono";
 import { bodyResultOf } from "../context/request";
-import { restoreSubscription$ } from "../services/zero-billing-restore.service";
+import { restoreSubscription$ } from "../services/billing-restore.service";
 import type { RouteEntry } from "../route-entry";
 
 const adminRequired = Object.freeze({
