+++ turbo/packages/db/scripts/agent-compose-consolidation-preflight-manifest.ts
@@ -205,6 +205,8 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/shared-thread.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/slack-message-context.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/slack-webhooks.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
+    "turbo/apps/api/src/signals/services/teams-connect.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
+    "turbo/apps/api/src/signals/services/teams-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/teams-dispatch.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/telegram-data.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/telegram-footer.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
@@ -226,8 +228,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/zero-slack-connect.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-slack-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-slack-data.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|@okouai/db/schema/agent-compose|agentComposeVersions:agentComposeVersions,agentComposes:agentComposes",
-    "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|@okouai/db/schema/agent-compose|agentComposes:agentComposes",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|@okouai/db/schema/zero-agent|zeroAgents:zeroAgents",
     "turbo/packages/db/src/schema/agent-draft.ts|./zero-agent|zeroAgents:zeroAgents",
@@ -334,6 +334,7 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/storage-volume-publication.service.ts|headVersionId",
     "turbo/apps/api/src/signals/services/storage-write.service.ts|headVersionId",
     "turbo/apps/api/src/signals/services/teams-chat-ingress.service.ts|agentComposeId",
+    "turbo/apps/api/src/signals/services/teams-connect.service.ts|agentComposeVersions,agentComposes,headVersionId,zeroAgents",
     "turbo/apps/api/src/signals/services/teams-dispatch.service.ts|agentComposeId,zeroAgents",
     "turbo/apps/api/src/signals/services/telegram-chat-ingress.service.ts|agentComposeId",
     "turbo/apps/api/src/signals/services/telegram-data.service.ts|agentComposeVersions,agentComposes,headVersionId",
@@ -354,7 +355,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "turbo/apps/api/src/signals/services/zero-runs-create.service.ts|agentComposeId,agentComposeVersions,agentComposes,headVersionId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-slack-connect.service.ts|agentComposes,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-slack-data.service.ts|zeroAgents",
-    "turbo/apps/api/src/signals/services/zero-teams-connect.service.ts|agentComposeVersions,agentComposes,headVersionId,zeroAgents",
     "turbo/apps/api/src/signals/services/zero-telegram-post.service.ts|agentComposeId,agentComposes,zeroAgents",
     "turbo/apps/platform/src/views/zero-page/zero-telegram-settings-page.tsx|headVersionId",
     "turbo/packages/api-contracts/src/contracts/composes.ts|headVersionId",
